Refactor pinot-avro module
