+++ pinot-server/src/test/java/org/apache/pinot/server/util/SegmentTestUtils.java
@@ -23,11 +23,11 @@ import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.apache.pinot.avro.data.readers.AvroUtils;
 import org.apache.pinot.spi.data.Schema;
 import org.apache.pinot.core.data.readers.FileFormat;
 import org.apache.pinot.core.indexsegment.generator.SegmentGeneratorConfig;
 import org.apache.pinot.core.indexsegment.generator.SegmentVersion;
-import org.apache.pinot.core.util.AvroUtils;
 
 
 public class SegmentTestUtils {
